requirements.txt
# Use PyTorch's official index for reliable CPU builds
--extra-index-url https://download.pytorch.org/whl/cpu

# Streamlit and core dependencies
streamlit
Pillow
requests

# PyTorch (A stable version for Python 3.11)
torch==2.3.1
torchvision
torchaudio

# Transformers and Hugging Face ecosystem
transformers
huggingface-hub
safetensors
tokenizers

# Explicitly pin TensorFlow to a modern, stable version
tensorflow==2.16.1

# PyTorch Lightning
pytorch-lightning

# Common numerical and utility libraries
pandas
PyYAML